feat: enable slices of maps on structs
Right now, the system only allows to  use `tags.*` for slices, but if
the slice is a complex type(struct or map) there is no way to validate.

This commits adds a way to test inside slice content if the value is a
map:

```
var data = map[string]interface{}{
  "user": "foo",
  "properties": []map[string]string{{
    "phone": "12001",
    "alias": "foobar",
  }},
}

v := validate.Map(data)

v.ConfigRules(map[string]string{
  "properties.*.admin": "bool",
  "properties.*.alias": "minlen:4",
})
```

Added tests inside a current test, I thought that makes sense to have in
there.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>